refactor: rename directory from trajectories->trajectory
