+++ source/state_representation/CMakeLists.txt
@@ -29,8 +29,8 @@ set(CORE_SOURCES
   src/parameters/Predicate.cpp
   src/geometry/Shape.cpp
   src/geometry/Ellipsoid.cpp
-  src/trajectories/CartesianTrajectory.cpp
-  src/trajectories/JointTrajectory.cpp
+  src/trajectory/CartesianTrajectory.cpp
+  src/trajectory/JointTrajectory.cpp
 )
 
 if (EXPERIMENTAL_FEATURES)
