+++ source/state_representation/include/state_representation/trajectory/TrajectoryBase.hpp
