+++ source/state_representation/test/tests/test_trajectory.cpp
@@ -1,7 +1,7 @@
 #include "state_representation/space/cartesian/CartesianState.hpp"
 #include "state_representation/space/joint/JointState.hpp"
-#include "state_representation/trajectories/CartesianTrajectory.hpp"
-#include "state_representation/trajectories/JointTrajectory.hpp"
+#include "state_representation/trajectory/CartesianTrajectory.hpp"
+#include "state_representation/trajectory/JointTrajectory.hpp"
 
 #include <eigen3/Eigen/src/Core/Matrix.h>
